Randomise calling of the control/candidate code
We want to call the control or candidate code in random order so we
don't have skewed results due to things like caching. So we wrap the
calling of the code in closures and call them in right order.

Test gives very generous margin of error, anything over 45% for each
method call passes.

Addresses Github issue #5.